+++ programs/drift/src/state/mod.rs
@@ -7,6 +7,7 @@ pub mod margin_calculation;
 pub mod oracle;
 pub mod oracle_map;
 pub mod order_params;
+pub mod paused_operations;
 pub mod perp_market;
 pub mod perp_market_map;
 pub mod spot_fulfillment_params;