+++ sdk/src/types.ts
@@ -27,10 +27,25 @@ export class MarketStatus {
 	static readonly DELISTED = { delisted: {} };
 }
 
+export enum PerpOperation {
+	UPDATE_FUNDING = 1,
+	AMM_FILL = 2,
+	FILL = 4,
+	SETTLE_PNL = 8,
+	SETTLE_PNL_WITH_POSITION = 16,
+}
+
+export enum SpotOperation {
+	UPDATE_CUMULATIVE_INTEREST = 1,
+	FILL = 2,
+	WITHDRAW = 4,
+}
+
 export enum UserStatus {
 	BEING_LIQUIDATED = 1,
 	BANKRUPT = 2,
 	REDUCE_ONLY = 4,
+	ADVANCED_LP = 8,
 }
 
 export class ContractType {
@@ -160,6 +175,9 @@ export class OrderActionExplanation {
 	static readonly REDUCE_ONLY_ORDER_INCREASED_POSITION = {
 		reduceOnlyOrderIncreasedPosition: {},
 	};
+	static readonly DERISK_LP = {
+		deriskLp: {},
+	};
 }
 
 export class OrderTriggerCondition {
@@ -183,6 +201,7 @@ export class DepositExplanation {
 	static readonly NONE = { none: {} };
 	static readonly TRANSFER = { transfer: {} };
 	static readonly BORROW = { borrow: {} };
+	static readonly REPAY_BORROW = { repayBorrow: {} };
 }
 
 export class SettlePnlExplanation {
@@ -343,6 +362,7 @@ export class LPAction {
 	static readonly ADD_LIQUIDITY = { addLiquidity: {} };
 	static readonly REMOVE_LIQUIDITY = { removeLiquidity: {} };
 	static readonly SETTLE_LIQUIDITY = { settleLiquidity: {} };
+	static readonly REMOVE_LIQUIDITY_DERISK = { removeLiquidityDerisk: {} };
 }
 
 export type FundingRateRecord = {
@@ -590,6 +610,7 @@ export type PerpMarketAccount = {
 	};
 	quoteSpotMarketIndex: number;
 	feeAdjustment: number;
+	pausedOperations: number;
 };
 
 export type HistoricalOracleData = {
@@ -681,6 +702,8 @@ export type SpotMarketAccount = {
 	flashLoanInitialTokenAmount: BN;
 
 	ordersEnabled: boolean;
+
+	pausedOperations: number;
 };
 
 export type PoolBalance = {