+++ CHANGELOG.md
@@ -8,15 +8,100 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Features
+
+### Fixes
+
+### Breaking
+
+## [2.61.0] - 2023-02-09
+
+### Features
+
+- program: better derivation of perp auction params when missing and for triggers ([#869](https://github.com/drift-labs/protocol-v2/pull/869))
+- program: calculate whether oracle's num quoters sufficient ([#860](https://github.com/drift-labs/protocol-v2/pull/860))
+
+### Fixes
+
+- program: include derisk lp order action explanation
+
+### Breaking
+
+## [2.60.0] - 2023-02-07
+
+### Features
+
+- program: sanitize perp auction params ([#859](https://github.com/drift-labs/protocol-v2/pull/859))
+- program: add repay borrow explanation ([#862](https://github.com/drift-labs/protocol-v2/pull/862))
+- program: derisk lp more granularly ([#849](https://github.com/drift-labs/protocol-v2/pull/849))
+
+### Fixes
+
+### Breaking
+
+## [2.59.0] - 2023-01-30
+
+### Features
+
+- program: separate out paused operations from market status ([#839](https://github.com/drift-labs/protocol-v2/pull/839))
+- program: use decayed last_oracle_conf_pct as lower bound for update ([#840](https://github.com/drift-labs/protocol-v2/pull/840))
+
+### Fixes
+
+### Breaking
+
+## [2.58.0] - 2023-01-27
+
+### Features
+
+### Fixes
+
+- program: AmmPaused doesnt block all fills
+
+### Breaking
+
+## [2.57.0] - 2023-01-25
+
+### Features
+
+- program: add recenter amm ix ([#836](https://github.com/drift-labs/protocol-v2/pull/836))
+
+### Fixes
+
+### Breaking
+
+## [2.56.0] - 2023-01-24
+
+### Features
+
+### Fixes
+
+- program: enable jit maker to fill same slot as taker placed ([#835](https://github.com/drift-labs/protocol-v2/pull/835))
+
+### Breaking
+
+## [2.55.0] - 2023-01-18
+
+### Features
+
+### Fixes
+
+- program: standardize lp shares in attempt_burn_user_lp_shares_for_risk_reduction ([#826](https://github.com/drift-labs/protocol-v2/pull/826))
+
+### Breaking
+
+## [2.54.0] - 2023-01-15
+
+### Features
+
 - sdk: move bracket orders into single instruction
 - sdk: add ability to do placeAndTake order with bracket orders attached
 - sdk: add option to cancel existing orders in market for place and take order
 - sdk: add option to get signed settlePnl tx back from a market order
-
 - program: auto derisk lp positions in settle pnl ([#766](https://github.com/drift-labs/protocol-v2/pull/766))
 - program: increase full perp liquidation threshold ([#807](https://github.com/drift-labs/protocol-v2/pull/807))
 - program: remove spot fee pool transfer ([#800](https://github.com/drift-labs/protocol-v2/pull/800))
 - program: increase insurance tier max ([#784](https://github.com/drift-labs/protocol-v2/pull/784))
+- sdk: can specify max custom margin ratio to initialize a new account with
 
 ### Fixes
 
@@ -24,7 +109,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ts-sdk: fix oracle is valid ([#806](https://github.com/drift-labs/protocol-v2/pull/806))
 
 ### Breaking
-
 
 ## [2.53.0] - 2023-12-31
 
@@ -43,7 +127,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
-- program: add ability to reclaim rent without deleting account ([#763](https://github.com/drift-labs/protocol-v2/pull/763)) 
+- program: add ability to reclaim rent without deleting account ([#763](https://github.com/drift-labs/protocol-v2/pull/763))
 - program: add borrow explanation to DepositRecords ([#772](https://github.com/drift-labs/protocol-v2/pull/772))
 - sdk: OrderSubscriber has resync option ([#780](https://github.com/drift-labs/protocol-v2/pull/780))
 - program: only consider recent last_active_slot in qualifies_for_withdraw_feen ([#756](https://github.com/drift-labs/protocol-v2/pull/756))
@@ -52,7 +136,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixes
 
 - program: handle underflow in calculate_liability_transfer_to_cover_margin_shortage ([#774](https://github.com/drift-labs/protocol-v2/pull/774))
-- program:  flip auction flag when trigger order adds auction ([#775](https://github.com/drift-labs/protocol-v2/pull/775))
+- program: flip auction flag when trigger order adds auction ([#775](https://github.com/drift-labs/protocol-v2/pull/775))
 - program: don't perform funding rate updates when slots_since_amm_update is stale ([#757](https://github.com/drift-labs/protocol-v2/pull/757))
 - program: add update last slot for filler in pay_keeper_flat_reward_for_spot